refactor(layout): rename files
